refactor: refactor all products page
