+++ src/modules/all-products/containers/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AllProductsContainer";
