+++ src/modules/all-products/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./AllProducts.hooks";
