+++ src/shared/hooks/useSelectApolloClientProduct/index.ts
@@ -1,0 +1,1 @@
+export * from "./useSelectApolloClientProduct";
